refactor: fake user 파일 위치 변경
